[Multiplayer] Fixed 200 warnings of type C4267 '+=': conversion from 'size_t' to 'int', possible loss of data.

+++ Descent3/multi_external.h
@@ -313,7 +313,7 @@ inline void MultiAddString(const char *str, uint8_t *data, int *count) {
 
   MultiAddByte((uint8_t)len, data, count);
   memcpy(&data[*count], str, len);
-  *count += len;
+  *count += static_cast<int>(len);
 }
 
 inline uint8_t MultiGetUbyte(uint8_t *data, int *count) {
